requirements.txt
SQLAlchemy~=1.4.42
SQLAlchemy-utils~=0.38.3
PyYAML~=6.0
bcrypt~=4.0.1
tornado~=6.2
requests~=2.28.1
beautifulsoup4~=4.11.1
prettytable~=3.4.1
